build: Default to use TI degenerate key for signing on non-HS devices
Switch to using the TI degenerate key to sign the system firmware binary
image on GP devices during the image tree blob build process. Using this
key speeds up the time needed by the ROM code execution for processing
this image and with this shortens our boot time.

The previous default behavior of signing with a random key can still
be triggered by passing an empty KEY variable to the make command as
follows:

$ make KEY=""

Signed-off-by: Andreas Dannenberg <dannenberg@ti.com>
Reviewed-by: Suman Anna <s-anna@ti.com>