Makefile: Fix HS builds by removing version reporting dependency
The rule for making the HS sysfw.bin file concatenates all its
dependencies, the rule for 'sysfw_version' is a phony target and does
not produce a file. Drop this dependency for HS.

While we are here lets do the same for non-HS builds. Phony targets
force the rebuilding of any rule that depends on them, this can
cause Make to perform more build steps than are actually needed.

Signed-off-by: Andrew F. Davis <afd@ti.com>
Tested-by: Denys Dmytriyenko <denys@ti.com>