j721e: rm-cfg: Start MAIN_0_R5_2 Tx and Rx at same level
Channel allocation for MAIN_0_R5_2 is done such that
Main NAV UDMA Tx and Rx channels start with different offsets.

This cannot be used for block copy carveout since the Tx and Rx channels
are not overlapping at start or end of the range.

Fix this by allocating more channels to MAIN_0_R5_0 so that the
MAIN_0_R5_2 channel ranges start at same value.

Autogen table takes care of the corresponding ring allocations as well.

Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>