HACK: j721e: rm-cfg: Use HOST_ALL for ethernet firmware flows
Ethernet firmware acts as server for providing networking functionality
to other clients. It allocates few flows from its pool for the client
and then retuns the same to the client.

Client does not own the resource it is borrowing from server and
currently SYSFW lacks any APIs to allow shared ownership of this
resource dyanmically.

In absence of such an API, mark the ethernet firmware flows are HOST_ALL
so that clients can call FLOW_CONFIG with the borrowed flow.

Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>