Makefile: Update firmware binaries to v2020.08-RC3
Update the Makefile to automatically fetch and build the v2020.08-RC3
version of the sysfw binaries for all supported SoCs from
ti-linux-firmware.

Note that this is the first version that will only support split
firmware architecture for j721e and j7200, so firmware with SciServer
enabled must be used with both SoCs.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>
Acked-by: Suman Anna <s-anna@ti.com>