soc: j721e: Update block copy allocation for UDMA channels
To use UDMA channels for block copy, the Tx and Rx channel number
has to be the same. When UDMA channels are allocated with just a
single range, sometimes it is not possible to allocate the ranges
such that the channels can be used for block copy usecase.

Fix this by allocating the channels in two ranges, first range for
block copy and second range for other usage. When there are no
channels for block copy, an entry with 0 count is added. This
is to maintain consistency when querying SYSFW about the
ranges allocated for a host.

Also adjust the MCU NAVSS interrupt router allocation after
the HSM re architecture

Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>