HACK: soc: j721e: Adjust MCU NAVSSS UDMA channel allocation
MCU R5 is facing hang when using channel no 5 and corresponding rings.
The transfer finishes, but the TR responce is not generated, ring
occupancy does not change.

Somehow, MCU R5 is having trouble to use channel no 5. UDMA works fine with
channel no 6 onwards.

As a workaround, increase the channel allocation for A72 and adjust
the allocation for all other cores.

Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>