scripts: gen_x509_combined_cert: Make DM boardcfg optional
Upon introduction of support for using DM firmware and providing DM
boardcfg through tiboot3.bin, support for providing ALL boardcfg as
part of one common boardcfg binary was lost. Modify the
gen_x509_combined_cert script so that the DM boardcfg can provided
optionally.

This allows all boardcfgs to be provided as part of a single binary to
look more like the traditional combined boot flow on platforms that do
not support a separate DM firmware.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>