soc: evm: am62ax: Allocate all DMASS1 chans
DMASS1 is the dedicated DMA subsystem for CSI-RX (cameras), and A53 is
the only user of CSI-RX on AM62a - so allocate all 6 channels to A53.

This is also required due to a bug in CSI-RX SHIM IP, where chan0 acts
as the default "dump" channel where frames not filtered by any other
channel end up. We want to avoid using chan0 for customer usecases, and
thus need all channels available for multi camera (4 channel) support.

Signed-off-by: Jai Luthra <j-luthra@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>